refactor(organizar_eventos): cambiar nombres de archivos para el gestor de eventos
